moving files to correct package, modifying _sqlite.py sync to fit refactor
