--- src/database_builder_libs/relational/__init__.py
@@ -1,7 +1,0 @@
-from database_builder_libs.relational._abstract import AbstractRelationalStore
-from database_builder_libs.relational._sqlite import SqliteRelationalStore
-
-__all__ = [
-    "AbstractRelationalStore",
-    "SqliteRelationalStore",
-]
