+++ src/database_builder_libs/utility/sync/_sqlite.py
@@ -4,7 +4,7 @@ from typing import Any, Optional
 
 from loguru import logger
 
-from database_builder_libs.relational._sqlite import SqliteRelationalStore
+from database_builder_libs.stores.sqlite._relational import SqliteRelationalStore
 from database_builder_libs.utility.sync._base import (
     AbstractSyncTracker,
     Artifact,
